requirements.txt
click
requests
pandas
python-decouple
tabulate
httpx
stringcase
textual
textual-dev
pyjwt
loguru
pydantic==2.10.6
pydantic[email]
fuzzywuzzy~=0.18.0
python-levenshtein<=0.26.1
aiofiles==24.1.0